Project Development Phase/sprint 2/backend/requirements.txt
alembic==1.8.1
autopep8==2.0.0
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
ibm-db==3.1.3
ibm-db-sa==0.3.8
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
pycodestyle==2.9.1
python-dotenv==0.21.0
python-http-client==3.3.7
sendgrid==6.9.7
SQLAlchemy==1.4.42
starkbank-ecdsa==2.2.0
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
